move Path to __init__
