+++ sphinxgallery/__init__.py
@@ -3,6 +3,47 @@
 import os
 __version__ = '0.0.10'
 
+
 def glr_path_static():
     """Returns path to packaged static files"""
     return os.path.abspath(os.path.join(os.path.dirname(__file__), '_static'))
+
+
+class Path(str):
+    """Path object for manipulating directory and file paths."""
+
+    def __init__(self, path):
+        super(Path, self).__init__(path)
+
+    @property
+    def isdir(self):
+        return os.path.isdir(self)
+
+    @property
+    def exists(self):
+        """Return True if path exists"""
+        return os.path.exists(self)
+
+    def pjoin(self, *args):
+        """Join paths. `p` prefix prevents confusion with string method."""
+        return self.__class__(os.path.join(self, *args))
+
+    def psplit(self):
+        """Split paths. `p` prefix prevents confusion with string method."""
+        return [self.__class__(p) for p in os.path.split(self)]
+
+    def makedirs(self):
+        if not self.exists:
+            os.makedirs(self)
+
+    def listdir(self):
+        return os.listdir(self)
+
+    def format(self, *args, **kwargs):
+        return self.__class__(super(Path, self).format(*args, **kwargs))
+
+    def __add__(self, other):
+        return self.__class__(super(Path, self).__add__(other))
+
+    def __iadd__(self, other):
+        return self.__add__(other)
