+++ sphinxgallery/gen_gallery.py
@@ -7,46 +7,7 @@ import os
 from . import glr_path_static
 from .gen_rst import generate_dir_rst
 from .docs_resolv import embed_code_links
-
-
-class Path(str):
-    """Path object for manipulating directory and file paths."""
-
-    def __init__(self, path):
-        super(Path, self).__init__(path)
-
-    @property
-    def isdir(self):
-        return os.path.isdir(self)
-
-    @property
-    def exists(self):
-        """Return True if path exists"""
-        return os.path.exists(self)
-
-    def pjoin(self, *args):
-        """Join paths. `p` prefix prevents confusion with string method."""
-        return self.__class__(os.path.join(self, *args))
-
-    def psplit(self):
-        """Split paths. `p` prefix prevents confusion with string method."""
-        return [self.__class__(p) for p in os.path.split(self)]
-
-    def makedirs(self):
-        if not self.exists:
-            os.makedirs(self)
-
-    def listdir(self):
-        return os.listdir(self)
-
-    def format(self, *args, **kwargs):
-        return self.__class__(super(Path, self).format(*args, **kwargs))
-
-    def __add__(self, other):
-        return self.__class__(super(Path, self).__add__(other))
-
-    def __iadd__(self, other):
-        return self.__add__(other)
+from . import Path
 
 
 def clean_gallery_out(build_dir):
