+++ sphinxgallery/tests/test_docs_resolv.py
@@ -5,6 +5,7 @@
 Testing the rst files generator
 """
 from __future__ import division, absolute_import, print_function
+from sphinxgallery import Path
 import sphinxgallery.docs_resolv as sg
 import tempfile
 import sys
@@ -15,7 +16,7 @@ def test_shelve():
     """Test if shelve can be caches information
     retrieved after file is deleted"""
     test_string = 'test information'
-    tmp_cache = tempfile.mkdtemp()
+    tmp_cache = Path(tempfile.mkdtemp())
     with tempfile.NamedTemporaryFile('w') as fid:
         fid.write(test_string)
         fid.seek(0)
